scripts/packet/requirements.txt
jsonschema>=4.17
pyyaml>=6.0.0
